chore: move console messages to their own module
