+++ src/index.js
@@ -1,6 +1,8 @@
 import printBuffer from './core';
 import { timer } from './helpers';
 import defaults from './defaults';
+import { REDUX_LOGGER_NOT_INSTALLED, REDUX_LOGGER_V3_BREAKING_CHANGE } from './messages';
+
 /* eslint max-len: ["error", 110, { "ignoreComments": true }] */
 /**
  * Creates logger with following options
@@ -41,23 +43,7 @@ function createLogger(options = {}) {
   // Detect if 'createLogger' was passed directly to 'applyMiddleware'.
   if (options.getState && options.dispatch) {
     // eslint-disable-next-line no-console
-    console.error(`[redux-logger] redux-logger not installed. Make sure to pass logger instance as middleware:
-// Logger with default options
-import { logger } from 'redux-logger'
-const store = createStore(
-  reducer,
-  applyMiddleware(logger)
-)
-// Or you can create your own logger with custom options http://bit.ly/redux-logger-options
-import { createLogger } from 'redux-logger'
-const logger = createLogger({
-  // ...options
-});
-const store = createStore(
-  reducer,
-  applyMiddleware(logger)
-)
-`);
+    console.error(REDUX_LOGGER_NOT_INSTALLED);
 
     return () => next => action => next(action);
   }
@@ -111,14 +97,7 @@ const defaultLogger = ({ dispatch, getState } = {}) => {
     return createLogger()({ dispatch, getState });
   }
   // eslint-disable-next-line no-console
-  console.error(`
-[redux-logger v3] BREAKING CHANGE
-[redux-logger v3] Since 3.0.0 redux-logger exports by default logger with default settings.
-[redux-logger v3] Change
-[redux-logger v3] import createLogger from 'redux-logger'
-[redux-logger v3] to
-[redux-logger v3] import { createLogger } from 'redux-logger'
-`);
+  console.error(REDUX_LOGGER_V3_BREAKING_CHANGE);
 };
 
 export { defaults, createLogger, defaultLogger as logger };
