+++ src/messages.js
@@ -1,0 +1,31 @@
+const REDUX_LOGGER_NOT_INSTALLED = `[redux-logger] redux-logger not installed. Make sure to pass logger instance as middleware:
+// Logger with default options
+import { logger } from 'redux-logger'
+const store = createStore(
+  reducer,
+  applyMiddleware(logger)
+)
+// Or you can create your own logger with custom options http://bit.ly/redux-logger-options
+import { createLogger } from 'redux-logger'
+const logger = createLogger({
+  // ...options
+});
+const store = createStore(
+  reducer,
+  applyMiddleware(logger)
+)
+`;
+
+const REDUX_LOGGER_V3_BREAKING_CHANGE = `
+[redux-logger v3] BREAKING CHANGE
+[redux-logger v3] Since 3.0.0 redux-logger exports by default logger with default settings.
+[redux-logger v3] Change
+[redux-logger v3] import createLogger from 'redux-logger'
+[redux-logger v3] to
+[redux-logger v3] import { createLogger } from 'redux-logger'
+`;
+
+export {
+  REDUX_LOGGER_NOT_INSTALLED,
+  REDUX_LOGGER_V3_BREAKING_CHANGE,
+};
